gcp/cloud_run_functions/requirements.txt
flask==3.1.0
functions-framework==3.8.2
google-genai==1.1.0
google-cloud-storage==3.0.0
google-cloud-firestore==2.3.0
